% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coregistration_images}
\alias{coregistration_images}
\title{Create a vector of the coregister images based on the images modalities}
\usage{
coregistration_images(vector)
}
\arguments{
\item{vector}{output object from the coregistration function.}
}
\description{
This function creates a vector of corregistered images for a patient.
The vector contains whether a vector of length one for only one modality (T2 or FLAIR)
or a vector of length two including both modalities (T2 and FLAIR).
}
