% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_bias_list}
\alias{create_bias_list}
\title{Create a vector of bias images based on the images modalities}
\usage{
create_bias_list(modalities, bias_T1, list_corregister)
}
\arguments{
\item{vector}{output object from the coregistration function.}
}
\description{
This function creates a vector of bias images for a patient.
The vector contains whether a vector of length one for only one modality (T1)
or a vector of length two/three including all modalities.
}
