% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_modalities}
\alias{get_modalities}
\title{Create a vector with the image modalities in each patient folder}
\usage{
get_modalities(patient)
}
\arguments{
\item{patient}{paths of MRI scan per patient.}
}
\description{
This function creates a vector with the name of the image modalitities fpr a patient.
}
