% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_mri_group}
\alias{load_mri_group}
\title{Load MRI per group or disease}
\usage{
load_mri_group(folder)
}
\arguments{
\item{folder}{folder containing the MRI scans.}
}
\value{
paths of MRI scans per patient if they exist.
}
\description{
This function loads the MRI scans from multiple patients. It assumes that the patients
folder with the the MRI scans are contained in a general folder.
}
