% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_modality_t1}
\alias{preprocess_modality_t1}
\title{Preprocesses T1-weighted MRI scan for one patient

This function preprocesses a raw T1-w MRI scan and generates a segmentation MRI scan using the FAST algorithm.
The preprocesising steps comprises imhomogeneity correction 'N4', registration to the MNI152 template with isotropic voxel size of 1mm^3
using the 'SyN' transformation, and skull stripping.}
\usage{
preprocess_modality_t1(
  mri.patient,
  folder.patient,
  atlas,
  mask,
  inhomogeneity = "N4",
  transformation = "SyN"
)
}
\arguments{
\item{mri.patient}{path of the T1-weighted scan.}

\item{folder.patient}{folder containing the T1-weighted scan. This folder usually refers to the patient.}

\item{atlas}{atlas template in NifTI format to spatially register the T1-weighted scans. By default the MNI152 atlas template is used.}

\item{mask}{brain mask in NifTI format of the atlas template to performed the skull stripping.}

\item{inhomogeneity}{inhomogeneity correction algorithm to be applied. The correction by default is the 'N4' bias correction.}

\item{transformation}{non-linear transformation for registering the T1-w MRI scan to the reference template. 'SyN' transformation is used by default.}
}
\value{
paths of preprocessed MRI scans.
}
\description{
Preprocesses T1-weighted MRI scan for one patient

This function preprocesses a raw T1-w MRI scan and generates a segmentation MRI scan using the FAST algorithm.
The preprocesising steps comprises imhomogeneity correction 'N4', registration to the MNI152 template with isotropic voxel size of 1mm^3
using the 'SyN' transformation, and skull stripping.
}
