% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_modalities}
\alias{get_modalities}
\title{Create a vector of strings with the MRI modalities available in a patient's folder.}
\usage{
get_modalities(patient)
}
\arguments{
\item{patient}{paths of MRI scans per patient.}
}
\value{
vector of strings containing the modalities available for a patient.
}
\description{
This function creates a vector with the name of the image modalitities for a patient.
}
