% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_modalities}
\alias{preprocess_modalities}
\title{Preprocess group of MRI scan for one patient}
\usage{
preprocess_modalities(
  mri.patient,
  folder.patient,
  modalities,
  atlas,
  mask,
  inhomogeneity = "N4",
  transformation = "SyN"
)
}
\arguments{
\item{mri.patient}{path of the MRI scans.}

\item{folder.patient}{folder containing the MRI scan. This folder usually refers as the patient.}

\item{modalities}{vector of modalities to be preprocessed. It must always contains the T1-w sequence scan.}

\item{atlas}{atlas template to spatially register the MRI scans. By default the MNI152 atlas template is used.}

\item{mask}{brain mask of the atlas template to performed the skull stripping.}

\item{inhomogeneity}{inhomogeneity correction algorithm to be applied. The correction by default is the 'N4' bias correction.}

\item{transformation}{non-linear transformation for registering the T1-w MRI scan to the reference template. 'SyN' transformation is used by default.}
}
\value{
paths of preprocessed MRI scans.
}
\description{
This function preprocesses raw T1-w, T2-w and FLAIR MRI scans and generates a segmentation MRI scan using the fast algorithm.'
The preprocesising steps comprises imhomogeneity correction 'N4', coregistration of other sequences to T1-w,
registration to the MNI152 template with isotropic voxel size of 1mm,
using the 'SyN' transformation, skull stripping, brain segmentation and intensity normalization via RAVEL or White Stripe.
}
