% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_modality_t1}
\alias{preprocess_modality_t1}
\title{Preprocess T1-w MRI scan for one patient}
\usage{
preprocess_modality_t1(
  mri_patient,
  folder_patient,
  atlas,
  mask,
  inhomogeneity = "N4",
  transformation = "SyN"
)
}
\arguments{
\item{mri_patient}{path of the T1-w scan.}

\item{folder_patient}{folder containing the T1-w scan. This folder usually refers as the patient.}

\item{atlas}{atlas template to spatially register the T1-w scans. By default the MNI152 atlas template is used.}

\item{mask}{brain mask of the atlas template to performed the skull stripping.}

\item{inhomogeneity}{inhomogeneity correction algorithm to be applied. The correction by default is the 'N4' bias correction.}

\item{trasnformation}{non-linear transformation for registering the T1-w MRI scan to the reference template. 'SyN' is used by default.}
}
\value{
paths of preprocessed MRI scans.
}
\description{
This function preprocesses a raw T1-w MRI scan and generates a spatially informed MRI data using the fast algorithm.'
The preprocesising steps comprises imhomogeneity correction 'N4', registration to the MNI152 template with voxel size of 2mm
using the 'SyN' transformation, skull stripping, and the RAVEL intensity normalization.
}
