% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_patients}
\alias{preprocess_patients}
\title{Preprocess MRI scans for multiple patients}
\usage{
preprocess_patients(patients.folder, clinical.covariates)
}
\arguments{
\item{clinical.covariates}{data.frame of covariates associated to the MRI scans. Number of rows should be equal to the number of images.}

\item{patients.general}{general folder containing sub-folders per patient with raw MRI images.}
}
\value{
paths of preprocessed MRI scans. MRI preprocessed images are stored in the patient's folder.
}
\description{
This function preprocesses raw T1-weighted, T2-weighted and/or FLAIR MRI scans and generates a brain segmentation MRI scans using the FAST algorithm.
The preprocessing steps comprise imhomogeneity correction 'N4', linear coregistration of T2-weighted and/or FLAIR to the T1-weighted, registration of all available modalities to the MNI152 template with an isotropic voxel size of 1mm^3
using the 'SyN' transformation, skull stripping, and RAVEL intensity normalization.
}
\examples{
# Get general folder
folder <- system.file("extdata", package = "neuronorm")
# Get covariates
covariates <- system.file("covariates.txt", package = "neuronorm")
# Read covariates information
clinical_info <- read.csv(file = covariates, sep = ';')
# Preprocess MRI scans: 'N4' inhomogeneity correction, 'SyN' non-linear transformation to MNI152 atlas template
# Brain extraction, Spatial informed MRI scan , a.k.a., brain segmentation and RAVEL intensity normalization only for T1-w images.
paths_preprocess_patients <- preprocess_patients(folder, clinical_info)
# Outputs paths of the preprocessed MRI scans.
paths_preprocess_patients$patient02$ravel
}
