% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_patients}
\alias{preprocess_patients}
\title{Preprocess T1-w MRI scan for multiple patients}
\usage{
preprocess_patients(patients.folder, clinical.covariates)
}
\arguments{
\item{patients.folder}{general folder containing folders per patient with raw T1-w images.}

\item{clinical.covariates}{table of covariates associated to the MRI scans. Number of rows should be equal to the number of images.}
}
\value{
paths of preprocessed MRI scans.
}
\description{
This function preprocesses raw T1-w MRI scans and generates a spatially informed MRI scans using the fast algorithm.
The preprocesising steps comprises imhomogeneity correction 'N4', registration to the MNI152 template with isotropic voxel size of 2mm
using the 'SyN' transformation, skull stripping, and RAVEL intensity normalization.
}
